Moved common methods to trait
